The P4 constants file can be more general, move it to common/lib.
